Implement multi-threaded multiplication of Sparse and Dense matrices
Unfortunately, I haven't figured out how to stack sparse and dense
matrices together, so it only works for Dense-Dense and Sparse-Sparse.
Going to fix ASAP, though.

Also, a new class was introduced: MatrixStacker. It just (what a surprise!)
stacks matrices vertically or horizontally.

And finally, there is an update to the Matrix interface: now it requires
one to also implement submatrix() method.